+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/main_method_examples/invalid/Name.java
@@ -2,11 +2,11 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.main_method_examples.invalid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class PublicStaticNoParam {
+public class Name {
 
-    public static void main() {
+    public static void myMain(String[] args) {
     }
 }
