+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/main_method_examples/invalid/ParameterCount.java
@@ -2,11 +2,12 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.invalid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.main_method_examples.invalid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class Name {
+public class ParameterCount {
 
-    public static void myMain(String[] args) {
+    static void main(String[] args, int count) {
+
     }
 }
