+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/main_method_examples/invalid/package-info.java
@@ -5,4 +5,4 @@
 /**
  * Contains invalid main method classes for the checker.
  */
-package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.invalid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.main_method_examples.invalid;
