+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/main_method_examples/valid/PackagePrivateInstanceNoParam.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.valid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.main_method_examples.valid;
 
 public class PackagePrivateInstanceNoParam {
