+++ src/test/java/de/uni_passau/fim/se2/pipeline_helper/checkers/main_method_examples/valid/PublicStaticNoParam.java
@@ -2,12 +2,11 @@
 //
 // SPDX-License-Identifier: CC0-1.0
 
-package de.uni_passau.fim.se2.pipeline_helper.checkers.mainMethodExamples.invalid;
+package de.uni_passau.fim.se2.pipeline_helper.checkers.main_method_examples.valid;
 
 @SuppressWarnings("checkstyle:HideUtilityClassConstructor")
-public class ParameterCount {
+public class PublicStaticNoParam {
 
-    static void main(String[] args, int count) {
-
+    public static void main() {
     }
 }
